+++ package.json
@@ -19,6 +19,7 @@
         "@types/ejs": "^2.7.0",
         "@types/express": "^4.16.0",
         "@types/lodash": "^4.14.123",
+        "@types/markdown-it": "0.0.9",
         "@types/mssql": "^4.0.12",
         "@types/mysql": "^2.15.5",
         "@types/node-fetch": "^2.1.4",
