+++ ts/index.ts
@@ -4,7 +4,7 @@ import * as bodyParser from 'body-parser';
 import cors from 'cors';
 import config from 'config';
 import { tableFromSql } from './db/mysql/tool';
-
+import { hello, markdown } from "./pages";
 const path = require('path');
 const fs = require('fs');
 const markdownIt = require('markdown-it'), md = new markdownIt();
@@ -46,31 +46,9 @@ const markdownIt = require('markdown-it'), md = new markdownIt();
     //设置模板引擎的格式即运用何种模板引擎
     app.set("view engine", "html");
 
-    app.use('/hello', async (req: Request, resp: Response) => {
-        let ret = await tableFromSql("select * from salestask$test.tv_mycustomer limit 10;");
-        let html = '<!DOCTYPE html><html lang = "en"> <head> <meta charset = "UTF-8" ><title> <%- title %> </title> <style type = "text/css" >h1 {color: blue;}</style></head><body> <div></div> <%- content %> </body></html>';
-        /**
-        ret.forEach(function (item) {
-            html += ejs.render('<div>dddd data <%= a %></div> <%- content %> <%= name %>', { a: 2, content: '<h1>Hello</h1>', name: item.mobile });
-        });
-        **/
-        let content = "# Hello I am markdown \n # 一级标题 \n ## 二级标题 \n ### 三级标题";
-        var mdresult = md.render(content);
-        let result = ejs.render(html, { title: '模板页面', content: mdresult })
-        resp.end(result);
-    });
+    app.use('/hello', hello);
 
-    app.use('/markdown', async (req: Request, resp: Response) => {
-        let html = '<!DOCTYPE html><html lang = "en"> <head> <meta charset = "UTF-8" ><title> <%- title %> </title> <style type = "text/css" >h1 {color: blue;}</style></head><body> <div></div> <%- content %> </body></html>';
-        fs.readFile("../contentmanagement/ts/dome.md", 'utf8', (err, data) => {
-            if (err) return;
-            else {
-                let mdresult = md.render(data);
-                let result = ejs.render(html, { title: '模板页面', content: mdresult });
-                resp.end(result);
-            }
-        })
-    });
+    app.use('/markdown', markdown);
 
     // 监听服务
     let port = config.get<number>('port');
