+++ ts/pages/index.ts
@@ -1,0 +1,6 @@
+export * from "./hello";
+export * from "./markdown";
+
+// export const pages = {
+//     "hello": hello,
+// }
